# Add `AppendFloat16ArrayToTensorProto` to acclerate `tf.constant` for float16