--- a/tensorflow/python/framework/fast_tensor_util.pyx
+++ b/tensorflow/python/framework/fast_tensor_util.pyx
@@ -11,7 +11,7 @@ def AppendFloat16ArrayToTensorProto(
     # For numpy, npy_half is a typedef for npy_uint16,
     # see: https://github.com/numpy/numpy/blob/master/doc/source/reference/c-api.coremath.rst#half-precision-functions
     # Because np.float16_t dosen't exist in cython, we use uint16_t here.
-    # TODO: use npy_half typedef?
+    # TODO: Use np.float16_t when cython supports it.
     tensor_proto, np.ndarray[np.uint16_t, ndim=1] nparray):
   cdef long i, n
   n = nparray.size