@facaiy 2018-05-15
The failure seems unrelated?

```python
======================================================================
261
FAIL: testPostPruningOfAllNodes (__main__.UpdateTreeEnsembleOpTest)
262
Test growing an ensemble with post-pruning, with all nodes are pruned.
263
----------------------------------------------------------------------
264
Traceback (most recent call last):
265
  File "T:/src/github/tensorflow/tensorflow/python/kernel_tests/boosted_trees/training_ops_test.py", line 1393, in testPostPruningOfAllNodes
266
    """, res_ensemble)
267
  File "C:\Python35\lib\site-packages\tensorflow\python\framework\test_util.py", line 881, in assertProtoEquals
268
    self._AssertProtoEquals(expected_message, message, msg=msg)
269
  File "C:\Python35\lib\site-packages\tensorflow\python\framework\test_util.py", line 858, in _AssertProtoEquals
270
    compare.assertProtoEqual(self, a, b, normalize_numbers=True, msg=msg)
271
  File "C:\Python35\lib\site-packages\tensorflow\python\util\protobuf\compare.py", line 107, in assertProtoEqual
272
    msg=msg)
273
AssertionError: 'tree[570 chars]4.0143\n  }\n}\ntree_metadata {\n}\ngrowing_me[88 chars]n}\n' != 'tree[570 chars]4.014299\n  }\n}\ntree_metadata {\n}\ngrowing_[90 chars]n}\n'
274
Diff is 849 characters long. Set self.maxDiff to None to see it. :
275
276
----------------------------------------------------------------------
```